git subrepo clone --branch=snippets-cache --force https://github.com/ccordoba12/python-lsp-server.git external-deps/python-lsp-server
subrepo:
  subdir:   "external-deps/python-lsp-server"
  merged:   "d92e90ae5"
upstream:
  origin:   "https://github.com/ccordoba12/python-lsp-server.git"
  branch:   "snippets-cache"
  commit:   "d92e90ae5"
git-subrepo:
  version:  "0.4.3"
  origin:   "https://github.com/ingydotnet/git-subrepo"
  commit:   "2f68596"
