+++ external-deps/python-lsp-server/.gitrepo
@@ -5,8 +5,8 @@
 ;
 [subrepo]
 	remote = https://github.com/python-lsp/python-lsp-server.git
-	branch = develop
-	commit = e8cf8e5b160f227550574cba480e276a98248d44
-	parent = b168e3445bbfbcb03aa04967668a2087a238f8a5
+	branch = snippets-cache
+	commit = d92e90ae5f688f663327d8e757f6c55c2f3320d8
+	parent = d6f3e21f2ed7d00fbd0a6642bb75b0021b55d6a1
 	method = merge
 	cmdver = 0.4.3
