+++ external-deps/python-lsp-server/pylsp/plugins/jedi_completion.py
@@ -3,13 +3,11 @@
 
 import logging
 import os.path as osp
-from collections import defaultdict
-from time import time
 
-from jedi.api.classes import Completion
 import parso
 
 from pylsp import _utils, hookimpl, lsp
+from pylsp.plugins._resolvers import LABEL_RESOLVER, SNIPPET_RESOLVER
 
 log = logging.getLogger(__name__)
 
@@ -57,10 +55,11 @@ def pylsp_completions(config, document, position):
     should_include_params = settings.get('include_params')
     should_include_class_objects = settings.get('include_class_objects', True)
 
-    max_labels_resolve = settings.get('resolve_at_most_labels', 25)
-    modules_to_cache_labels_for = settings.get('cache_labels_for', None)
-    if modules_to_cache_labels_for is not None:
-        LABEL_RESOLVER.cached_modules = modules_to_cache_labels_for
+    max_to_resolve = settings.get('resolve_at_most', 25)
+    modules_to_cache_for = settings.get('cache_for', None)
+    if modules_to_cache_for is not None:
+        LABEL_RESOLVER.cached_modules = modules_to_cache_for
+        SNIPPET_RESOLVER.cached_modules = modules_to_cache_for
 
     include_params = snippet_support and should_include_params and use_snippets(document, position)
     include_class_objects = snippet_support and should_include_class_objects and use_snippets(document, position)
@@ -70,7 +69,7 @@ def pylsp_completions(config, document, position):
             c,
             include_params,
             resolve=resolve_eagerly,
-            resolve_label=(i < max_labels_resolve)
+            resolve_label_or_snippet=(i < max_to_resolve)
         )
         for i, c in enumerate(completions)
     ]
@@ -83,7 +82,7 @@ def pylsp_completions(config, document, position):
                     c,
                     False,
                     resolve=resolve_eagerly,
-                    resolve_label=(i < max_labels_resolve)
+                    resolve_label_or_snippet=(i < max_to_resolve)
                 )
                 completion_dict['kind'] = lsp.CompletionItemKind.TypeParameter
                 completion_dict['label'] += ' object'
@@ -175,9 +174,9 @@ def _resolve_completion(completion, d):
     return completion
 
 
-def _format_completion(d, include_params=True, resolve=False, resolve_label=False):
+def _format_completion(d, include_params=True, resolve=False, resolve_label_or_snippet=False):
     completion = {
-        'label': _label(d, resolve_label),
+        'label': _label(d, resolve_label_or_snippet),
         'kind': _TYPE_MAP.get(d.type),
         'sortText': _sort_text(d),
         'insertText': d.name
@@ -193,29 +192,8 @@ def _format_completion(d, include_params=True, resolve=False, resolve_label=Fals
         completion['insertText'] = path
 
     if include_params and not is_exception_class(d.name):
-        sig = d.get_signatures()
-        if not sig:
-            return completion
-
-        positional_args = [param for param in sig[0].params
-                           if '=' not in param.description and
-                           param.name not in {'/', '*'}]
-
-        if len(positional_args) > 1:
-            # For completions with params, we can generate a snippet instead
-            completion['insertTextFormat'] = lsp.InsertTextFormat.Snippet
-            snippet = d.name + '('
-            for i, param in enumerate(positional_args):
-                snippet += '${%s:%s}' % (i + 1, param.name)
-                if i < len(positional_args) - 1:
-                    snippet += ', '
-            snippet += ')$0'
-            completion['insertText'] = snippet
-        elif len(positional_args) == 1:
-            completion['insertTextFormat'] = lsp.InsertTextFormat.Snippet
-            completion['insertText'] = d.name + '($0)'
-        else:
-            completion['insertText'] = d.name + '()'
+        snippet = _snippet(d, resolve_label_or_snippet)
+        completion.update(snippet)
 
     return completion
 
@@ -227,6 +205,13 @@ def _label(definition, resolve=False):
     if sig:
         return sig
     return definition.name
+
+
+def _snippet(definition, resolve=False):
+    if not resolve:
+        return {}
+    snippet = SNIPPET_RESOLVER.get_or_create(definition)
+    return snippet
 
 
 def _detail(definition):
@@ -244,86 +229,3 @@ def _sort_text(definition):
     # If its 'hidden', put it next last
     prefix = 'z{}' if definition.name.startswith('_') else 'a{}'
     return prefix.format(definition.name)
-
-
-class LabelResolver:
-
-    def __init__(self, format_label_callback, time_to_live=60 * 30):
-        self.format_label = format_label_callback
-        self._cache = {}
-        self._time_to_live = time_to_live
-        self._cache_ttl = defaultdict(set)
-        self._clear_every = 2
-        # see https://github.com/davidhalter/jedi/blob/master/jedi/inference/helpers.py#L194-L202
-        self._cached_modules = {'pandas', 'numpy', 'tensorflow', 'matplotlib'}
-
-    @property
-    def cached_modules(self):
-        return self._cached_modules
-
-    @cached_modules.setter
-    def cached_modules(self, new_value):
-        self._cached_modules = set(new_value)
-
-    def clear_outdated(self):
-        now = self.time_key()
-        to_clear = [
-            timestamp
-            for timestamp in self._cache_ttl
-            if timestamp < now
-        ]
-        for time_key in to_clear:
-            for key in self._cache_ttl[time_key]:
-                del self._cache[key]
-            del self._cache_ttl[time_key]
-
-    def time_key(self):
-        return int(time() / self._time_to_live)
-
-    def get_or_create(self, completion: Completion):
-        if not completion.full_name:
-            use_cache = False
-        else:
-            module_parts = completion.full_name.split('.')
-            use_cache = module_parts and module_parts[0] in self._cached_modules
-
-        if use_cache:
-            key = self._create_completion_id(completion)
-            if key not in self._cache:
-                if self.time_key() % self._clear_every == 0:
-                    self.clear_outdated()
-
-                self._cache[key] = self.resolve_label(completion)
-                self._cache_ttl[self.time_key()].add(key)
-            return self._cache[key]
-
-        return self.resolve_label(completion)
-
-    def _create_completion_id(self, completion: Completion):
-        return (
-            completion.full_name, completion.module_path,
-            completion.line, completion.column,
-            self.time_key()
-        )
-
-    def resolve_label(self, completion):
-        try:
-            sig = completion.get_signatures()
-            return self.format_label(completion, sig)
-        except Exception as e:  # pylint: disable=broad-except
-            log.warning(
-                'Something went wrong when resolving label for {completion}: {e}',
-                completion=completion, e=e
-            )
-            return ''
-
-
-def format_label(completion, sig):
-    if sig and completion.type in ('function', 'method'):
-        params = ', '.join(param.name for param in sig[0].params)
-        label = '{}({})'.format(completion.name, params)
-        return label
-    return completion.name
-
-
-LABEL_RESOLVER = LabelResolver(format_label)
